Move fishing-scene tests
