+++ svelte/src/scenes/fishing/__tests__/fishing-scene.cy.ts
@@ -1,16 +1,13 @@
-import { makeFishingScene } from "../fishing/fishing-scene";
-import ViewSceneFixture from "./ViewSceneFixture.svelte";
+import { makeFishingScene } from "../fishing-scene";
 
 describe("fishing scene", () => {
   describe("rendering tests", () => {
     beforeEach(() => {
       cy.viewport(1400, 900);
 
-      cy.mount(ViewSceneFixture, {
-        props: {
-          initialSceneSpec: makeFishingScene,
-          seed: "abcd",
-        },
+      cy.mountViewScene({
+        initialSceneSpec: makeFishingScene,
+        seed: "abcd",
       });
       cy.get("canvas").should("have.attr", "data-initialised", "true");
     });
